sync with the code from the source repo
We were mostly on target with the sample code, only a few places were different and mostly that was asking for write permissions when we only needed read.

There was a bug in the monster movement that did get fixed and would have been nearly impossible to see without this sync.

This also brings in the mouse support I didn't initially want to implement, but it's how they handle ranged combat actions later on so I circled back around.